@@ -7,10 +7,9 @@
     <artifactId>distancemeter</artifactId>
     <version>1.0-SNAPSHOT</version>
     <properties>
-        <compiler-plugin.version>3.14.0</compiler-plugin.version>
-
-        <maven.compiler.release>17</maven.compiler.release>
-        <javac.target>17</javac.target>
+        <maven.compiler.parameters>true</maven.compiler.parameters>
+        <maven.compiler.source>21</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
